split serializer to multiple files
