+++ packages/squiggle-lang/tsconfig.build.json
@@ -6,5 +6,6 @@
     "outDir": "dist",
     "tsBuildInfoFile": "dist/tsconfig.build.tsbuildinfo"
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "references": [{ "path": "../serializer" }]
 }
